Moving Agents SPA out of admin namespace. * Making Agents SPA the default view. * Remove agents SPA toggle. * Disable all buttons for operations for non admins. Non admins can only view the agents page. * Remove checkboxes for non-admins in agents page. * Do not display job run history link for non admin user.
