+++ server/resources/available.toggles
@@ -10,11 +10,6 @@
          "key": "pipeline_config_single_page_app_key",
          "description": "Enables the pipeline config single page application. Default: enabled.",
          "value": true
-      },
-      {
-        "key": "agents_single_page_app_key",
-        "description": "Enables the agents single page application. Default: enabled.",
-        "value": true
       }
    ]
 }
