+++ server/webapp/WEB-INF/rails.new/app/assets/new_javascripts/views/agents/agent_row_widget.js.msx
@@ -63,20 +63,33 @@ define(['mithril', 'lodash', 'helpers/form_helper', 'views/agents/build_details_
       var environments    = joinOrNoneSpecified(agent.environments());
       var agentStatus     = agent.status();
       var isBuildingAgent = !agent.buildDetails().isEmpty();
+      var selectAgentCheckbox;
+      var hostnameLink = <span>{agent.hostname()}</span>;
 
       if (isBuildingAgent) {
         agentStatus = <f.link onclick={ctrl.buildDetailsClicked.bind(ctrl, agent.uuid())}
                               class="has-build-details-drop-down">{agent.status()}</f.link>;
       }
 
+      if (args.isUserAdmin) {
+        selectAgentCheckbox = (
+          <input type="checkbox"
+                 checked={args.checkBoxModel()}
+                 onclick={m.withAttr('checked', args.checkBoxModel)}/>
+        );
+        hostnameLink = (
+          <a href={agentJobHistoryPath(agent.uuid())}>{agent.hostname()}</a>
+        );
+      }
+
       return (
         <tr key={agent.uuid()} class={agent.status().toLowerCase()}>
           <td key="checkbox">
-            <input type="checkbox"
-                   checked={args.checkBoxModel()}
-                   onclick={m.withAttr('checked', args.checkBoxModel)}/>
+            {selectAgentCheckbox}
           </td>
-          <td key="hostname"><a href={agentJobHistoryPath(agent.uuid())}>{agent.hostname()}</a></td>
+          <td key="hostname">
+            {hostnameLink}
+          </td>
           <td key="sandbox">{agent.sandbox()}</td>
           <td key="operatingSystem">{agent.operatingSystem()}</td>
           <td key="ipAddress">{agent.ipAddress()}</td>
