+++ server/webapp/WEB-INF/rails.new/app/assets/stylesheets/module.scss
@@ -156,11 +156,11 @@
 /*
 * Agents
 */
-#agents .entity_title {
+#old_agents .entity_title {
     float: left;
 }
 
-#agents .content_wrapper_outer {
+#old_agents .content_wrapper_outer {
     clear: both;
 }
