+++ server/webapp/WEB-INF/rails.new/app/assets/stylesheets/views/old_agents.scss
@@ -1,23 +1,23 @@
-/*************************GO-LICENSE-START*********************************
- * Copyright 2014 ThoughtWorks, Inc.
+/*!
+ * Copyright 2016 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *************************GO-LICENSE-END***********************************/
+ */
 
 /**
  * Agents.css
  */
- 
+
 
 /**
  * TABLE OF CONTENTS
@@ -33,72 +33,72 @@
  * =INDEX
 **/
 #agents_form_container {
- position: relative;
- top: 0;
- left: 0;
+  position: relative;
+  top: 0;
+  left: 0;
 }
 
 #agents_form {
-    border-right:1px solid #d0d0d0;;
-    padding-right: 10px;
-    margin-right: 10px;
+  border-right:1px solid #d0d0d0;;
+  padding-right: 10px;
+  margin-right: 10px;
 }
 
 #agents_form_container .edit_panel {
-    padding: 0;
-    margin-bottom: 10px;
-    background: #FFF;
-    -moz-border-radius: 0px;
-    -webkit-border-radius: 0px;
-    border-radius: 0px;
-    border: 0;
-    padding: 0 0;
+  padding: 0;
+  margin-bottom: 10px;
+  background: #FFF;
+  -moz-border-radius: 0px;
+  -webkit-border-radius: 0px;
+  border-radius: 0px;
+  border: 0;
+  padding: 0 0;
 }
 
 .page_header {
   min-height: 0;
 }
 
-#agents .filter_agents {
-    float: right;
+#old_agents .filter_agents {
+  float: right;
 }
 
-#agents #agents_filter_form .heading {
-    font-weight: bold;
+#old_agents #agents_filter_form .heading {
+  font-weight: bold;
 }
 #filter_help{
-    max-width: 448px;
+  max-width: 448px;
 }
 
-#agents .filter_agents #clear_filter {
-    float: none;
+#old_agents .filter_agents #clear_filter {
+  float: none;
 }
 
 #filter_help .filter_help_instructions {
-    width:228px;
-    margin :10px;
-    font-size:11px;
+  width:228px;
+  margin :10px;
+  font-size:11px;
 }
 
 #filter_help .filter_help_instructions a{
-    font-size:11px;
-    color:#1F66BD;
+  font-size:11px;
+  color:#1F66BD;
 }
 
 #filter_help .filter_help_instructions p{
-    padding-bottom:6px;
+  padding-bottom:6px;
 }
 
 /* Docked Header Styles */
 
 #agents_form_container .isFixed {
-    background: image_url("g9/backgrounds/bg_actionpanel_shadow.png") repeat-x scroll 0 0 transparent;
-    padding: 8px 15px;
-    position: fixed;
-    top: 87px;
-    right: 0;
-    z-index: 100000;
-    height: 35px;
+  background: image_url("g9/backgrounds/bg_actionpanel_shadow.png") repeat-x scroll 0 0 transparent;
+  padding: 8px 15px;
+  position: fixed;
+  top: 87px;
+  right: 0;
+  z-index: 100000;
+  height: 35px;
 }
 
 .isFixed .download_agent {
@@ -107,18 +107,18 @@
 }
 
 .page_header.isFixed {
-    position: fixed;
-    top:0;
-    z-index: 1000;
-    width: 98%;
+  position: fixed;
+  top:0;
+  z-index: 1000;
+  width: 98%;
 }
 
 .filter_agents.isDocked {
-    margin-right: 3px;
+  margin-right: 3px;
 }
 
 .page_header.isFixed .filter_agents {
-    padding-right: 5px;
+  padding-right: 5px;
 }
 
 #dd_ajax_float {
@@ -131,20 +131,20 @@
   * =SHOW
 **/
 .agent_details_pane {
-    margin-top: 25px;
+  margin-top: 25px;
 }
 .agent_details_pane label {
-    color: #828282;
-    float: left;
-    width: 140px;
+  color: #828282;
+  float: left;
+  width: 140px;
 }
 .agent_details_pane div {
-    margin-bottom: 10px;
+  margin-bottom: 10px;
 }
 
 .agent_header th.selector {
-    padding: 0 2px 0 5px;
-    text-align: center;
+  padding: 0 2px 0 5px;
+  text-align: center;
 }
 
 /**
@@ -167,26 +167,26 @@
 **/
 
 .agent_header th.hostname {
-    padding-left: 34px;
+  padding-left: 34px;
 }
 
 .bootstraper_version_warning {
-    margin: 17px 0px 20px;
+  margin: 17px 0px 20px;
 }
 
 .agent_instance .agent_hostname {
-    padding-left: 25px;
-    line-height: 16px;
-    height: 16px;
-    display: block;
+  padding-left: 25px;
+  line-height: 16px;
+  height: 16px;
+  display: block;
 }
 
 .agent_instance .agent_hostname.bootstrap-old {
-    background: imageurl('g9/icons/icon_info_16.png') no-repeat scroll 0 0 transparent;
+  background: imageurl('g9/icons/icon_info_16.png') no-repeat scroll 0 0 transparent;
 }
 
 .agent_instance .info_icon {
-    margin-bottom: -3px;
+  margin-bottom: -3px;
 }
