--- server/webapp/WEB-INF/rails.new/app/views/admin/agents/index.html.erb
@@ -1,3 +1,0 @@
-<div id="agents">
-  <span class="page-spinner"></span>
-</div>
