--- server/webapp/WEB-INF/rails.new/app/views/agents/index.json.erb
@@ -1,5 +1,0 @@
-{
-    "actual_agent_selectors" : {"html" : <%== render_json(:partial => "agents/hidden_selectors.html", :locals => {:scope => {:agents => @agents}}) -%>},
-    "ajax_agents_header" : {"html" : <%== render_json(:partial => 'agents_header.html.erb', :locals => {:scope => {}}) %>},
-    "ajax_agents_table" : {"html" : <%== render_json(:partial => 'agents_table.html.erb', :locals => {:scope => {}}) %>}
-}
