+++ server/webapp/WEB-INF/rails.new/app/views/layouts/agent_detail.html.erb
@@ -16,7 +16,7 @@
         </div>
         <div class="page_header">
             <ul class="entity_title">
-                <!--<li class="name"><%= link_to l.string("AGENTS_POPUP_TITLE"), url_for(:controller => "agents", :action => "index") -%> </li>-->
+                <!--<li class="name"><%= link_to l.string("AGENTS_POPUP_TITLE"), url_for(:controller => "old_agents", :action => "index") -%> </li>-->
                 <li class="last"><h1><%= @agent.getHostname() -%></h1></li>
             </ul>
