+++ server/webapp/WEB-INF/rails.new/app/views/old_agents/_agents_table.html.erb
