+++ server/webapp/WEB-INF/rails.new/app/views/old_agents/_hidden_selectors.html.erb
