+++ server/webapp/WEB-INF/rails.new/app/views/old_agents/index.json.erb
@@ -1,0 +1,5 @@
+{
+    "actual_agent_selectors" : {"html" : <%== render_json(:partial => "old_agents/hidden_selectors.html", :locals => {:scope => {:agents => @agents}}) -%>},
+    "ajax_agents_header" : {"html" : <%== render_json(:partial => 'old_agents/agents_header.html.erb', :locals => {:scope => {}}) %>},
+    "ajax_agents_table" : {"html" : <%== render_json(:partial => 'old_agents/agents_table.html.erb', :locals => {:scope => {}}) %>}
+}
