+++ server/webapp/WEB-INF/rails.new/app/views/old_agents/resource_selector.html.erb
