+++ server/webapp/WEB-INF/urlrewrite.xml
@@ -75,7 +75,7 @@
             Prepends 'rails/' to Rails urls, so that the request does not go through Java at all
             (currently RackFilter passes a request to Rails only if Java returns a 40* HTTP response
         </note>
-        <from>^/((jasmine-specs)|(jasmine-specs/*)|(jasmine-specs-new)|(jasmine-specs-new/*)|(about)|(agents)|(agents/.*)|(agents\..*)|(environments)|(environments/.*)|(environments\..*)|(pipelines/.*)|(pipelines\.json)|(materials/.*)|(home/?)|(pipelines)|(history/.*)|(shine)|(shine/.*)|(gadgets/.*)|(server)|(server/.*)|(admin/config/.*)|(admin/users)|(admin/users/.*)|(admin/oauth/.*)|(admin/gadgets/.*)|(oauth/.*)|(gadgets/.*)|(failures/.*)|(compare/.*)|(admin/agents)|(admin/pipelines)|(admin/pipelines/.*)|(admin/pipeline/.*)|(admin/pipeline_group)|(admin/pipeline_group/.*)|(admin/templates)|(admin/templates/.*)|(admin/backup)|(admin/backup/.*)|(admin/commands)|(admin/commands/.*)|(admin/plugins)|(admin/plugins/.*)|(config_change/.*)|(admin/config_xml)|(admin/config_xml/.*)|(admin/package_repositories)|(admin/package_repositories/.*)|(config_view/.*)|(admin/package_definitions)|(admin/package_definitions/.*)|(admin/materials)|(admin/materials/.*)|(admin/garage)|(admin/garage/.*)|(assets/.*)|(cctray.xml)|(errors/inactive))$</from>
+        <from>^/((jasmine-specs)|(jasmine-specs/*)|(jasmine-specs-new)|(jasmine-specs-new/*)|(about)|(agents)|(old_agents)|(old_agents/.*)|(old_agents\..*)|(agents/.*)|(agents\..*)|(environments)|(environments/.*)|(environments\..*)|(pipelines/.*)|(pipelines\.json)|(materials/.*)|(home/?)|(pipelines)|(history/.*)|(shine)|(shine/.*)|(gadgets/.*)|(server)|(server/.*)|(admin/config/.*)|(admin/users)|(admin/users/.*)|(admin/oauth/.*)|(admin/gadgets/.*)|(oauth/.*)|(gadgets/.*)|(failures/.*)|(compare/.*)|(admin/agents)|(admin/pipelines)|(admin/pipelines/.*)|(admin/pipeline/.*)|(admin/pipeline_group)|(admin/pipeline_group/.*)|(admin/templates)|(admin/templates/.*)|(admin/backup)|(admin/backup/.*)|(admin/commands)|(admin/commands/.*)|(admin/plugins)|(admin/plugins/.*)|(config_change/.*)|(admin/config_xml)|(admin/config_xml/.*)|(admin/package_repositories)|(admin/package_repositories/.*)|(config_view/.*)|(admin/package_definitions)|(admin/package_definitions/.*)|(admin/materials)|(admin/materials/.*)|(admin/garage)|(admin/garage/.*)|(assets/.*)|(cctray.xml)|(errors/inactive))$</from>
         <to last="true">/rails/$1</to>
         <set name="rails_bound">true</set>
     </rule>
